# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Teleassistance API Documentation",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/alerts": {
            "get": {
                "tags": [
                    "Alerts"
                ],
                "summary": "Muestra todas las alertas paginadas",
                "operationId": "4478ffd7354aa9ca6132fb9e74bdde0b",
                "responses": {
                    "200": {
                        "description": "Lista de alertas",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/AlertResource"
                                            }
                                        },
                                        "links": {
                                            "type": "object"
                                        },
                                        "meta": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Alerts"
                ],
                "summary": "Crea una nueva alerta",
                "operationId": "42b5297a4f4f281e294bf8339a0a80d8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreAlertRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Alerta creada con éxito",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AlertResource"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/alerts/{id}": {
            "get": {
                "tags": [
                    "Alerts"
                ],
                "summary": "Muestra una alerta específica",
                "operationId": "48e65a8f03a14c0b36563c5284e726a3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la alerta",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Alerta mostrada con éxito",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AlertResource"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Alerts"
                ],
                "summary": "Actualiza una alerta",
                "operationId": "037adc62bb0d80e7c7a153cf2445e337",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la alerta",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateAlertRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Alerta actualizada con éxito",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AlertResource"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Alerts"
                ],
                "summary": "Elimina una alerta",
                "operationId": "6dea41ff5a0bbf10775b84b2275d50e0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la alerta",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Alerta eliminada con éxito"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "StoreAlertRequest": {
                "description": "Validación para la creación de alertas",
                "required": [
                    "alert_subtype_id",
                    "title",
                    "description",
                    "frequency",
                    "zone_id"
                ],
                "properties": {
                    "alert_subtype_id": {
                        "description": "ID del subtipo de alerta",
                        "type": "integer",
                        "example": 1
                    },
                    "title": {
                        "description": "Título de la alerta",
                        "type": "string",
                        "maxLength": 255,
                        "example": "Alerta de emergencia"
                    },
                    "description": {
                        "description": "Descripción detallada de la alerta",
                        "type": "string",
                        "example": "Se ha detectado una anomalía en la zona asignada"
                    },
                    "frequency": {
                        "description": "Frecuencia de la alerta",
                        "type": "string",
                        "enum": [
                            "Puntual",
                            "Diaria",
                            "Varios días",
                            "Semanal",
                            "Mensual"
                        ],
                        "example": "Diaria"
                    },
                    "days_of_week": {
                        "description": "Días de la semana en formato JSON (opcional, solo si aplica)",
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "nullable": true
                    },
                    "zone_id": {
                        "description": "ID de la zona donde se generó la alerta",
                        "type": "integer",
                        "example": 3
                    }
                },
                "type": "object"
            },
...
            "UpdateAlertRequest": {
                "description": "Validación para la actualización de alertas",
                "required": [
                    "alert_subtype_id",
                    "title",
                    "description",
                    "frequency",
                    "zone_id"
                ],
                "properties": {
                    "alert_subtype_id": {
                        "description": "ID del subtipo de alerta",
                        "type": "integer",
                        "example": 3
                    },
                    "title": {
                        "description": "Título de la alerta",
                        "type": "string",
                        "maxLength": 255,
                        "example": "Alerta de emergencia"
                    },
                    "description": {
                        "description": "Descripción detallada de la alerta",
                        "type": "string",
                        "example": "El paciente ha activado el botón de emergencia."
                    },
                    "frequency": {
                        "description": "Frecuencia de la alerta",
                        "type": "string",
                        "enum": [
                            "Puntual",
                            "Diaria",
                            "Varios días",
                            "Semanal",
                            "Mensual"
                        ],
                        "example": "Diaria"
                    },
                    "days_of_week": {
                        "description": "Días de la semana en los que se activa la alerta (JSON válido)",
                        "type": "string",
                        "nullable": true
                    },
                    "zone_id": {
                        "description": "ID de la zona asociada a la alerta",
                        "type": "integer",
                        "example": 2
                    }
                },
                "type": "object"
            },
...
            "AlertResource": {
                "description": "Esquema del recurso Alerta",
                "properties": {
                    "id": {
                        "description": "Identificador de la alerta",
                        "type": "integer"
                    },
                    "alert": {
                        "$ref": "#/components/schemas/AlertSubtypeCompactResource"
                    },
                    "frequency": {
                        "description": "Frecuencia de la alerta",
                        "type": "string"
                    },
                    "days_of_week": {
                        "description": "Días de la semana en los que se activa la alerta",
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "zone": {
                        "$ref": "#/components/schemas/ZoneResource"
                    }
                },
                "type": "object"
            },
            "AlertSubtypeCompactResource": {
                "description": "Esquema del recurso Subtipo de Alerta",
                "properties": {
                    "name": {
                        "description": "Nombre del tipo de alerta",
                        "type": "string"
                    },
                    "description": {
                        "description": "Descripción del subtipo de alerta",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "ZoneResource": {
                "description": "Esquema del recurso Zona",
                "properties": {
                    "id": {
                        "description": "Identificador de la zona",
                        "type": "integer"
                    },
                    "city": {
                        "description": "Nombre de la ciudad",
                        "type": "string"
                    },
                    "zone": {
                        "description": "Nombre de la zona dentro de la ciudad",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}